requirements.txt
clearml==1.18.0
datasets==3.6.0
hydra-core==1.3.2
implicit==0.7.2
lightfm==1.17
numpy==2.2.4
omegaconf==2.3.0
pandas==2.2.2
polars==1.31.0
pyarrow==20.0.0
pytorch-lifestream==0.6.0
pytorch-lightning==2.5.1
replay-rec==0.18.1
scikit-learn==1.6.1
scipy==1.15.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
vllm==0.9.1
